requirements.txt
amqp==5.1.1
argh==0.26.2
asgiref==3.5.0
async-timeout==4.0.3
billiard==4.1.0
celery==5.3.1
click==8.1.7
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.3.0
colorama==0.4.6
config2py==0.0.3
coverage==6.3.1
cron-descriptor==1.4.0
Django==4.0.2
django-celery-beat==2.5.0
django-cors-headers==4.2.0
django-extensions==3.1.5
django-jazzmin==2.4.9
django-mathfilters==1.0.0
django-sass-processor==1.1
django-timezone-field==6.0.1
djangorestframework==3.13.1
djangorestframework-recursive==0.1.2
djangorestframework-simplejwt==5.3.0
dol==0.1.71
gunicorn==20.1.0
importlib-resources==5.4.0
kombu==5.3.2
prompt-toolkit==3.0.39
psycopg2-binary==2.9.7
py2store==0.1.11
PyJWT==2.8.0
python-crontab==3.0.0
python-dateutil==2.8.2
pytz==2021.3
redis==5.0.0
rest-framework-simplejwt==0.0.2
six==1.16.0
sqlparse==0.4.2
tzdata==2023.3
vine==5.0.0
wcwidth==0.2.6